populated this moose project from prompt in cursor
I'm working on building my first moose project.
@https://docs.fiveonefour.com/moose

Help me build my moose project!
In the project here so far, I have a python boilerplat project for moose.

The ingest pipeline has the Baz, Foo, and Bar models...
these can dissappear they're not required for my project.

Instead we should define data models for the data we want!
This analytics project needs to be able to ingest data from websites that contain set lists.
To start, let's do Goose set lists. Later, we'll want to expand to other bands. So we should leave ourselves open for extension.
@https://elgoose.net/setlists/
@https://elgoose.net/song/draconian-meter-ma

For each show, the set lists should be recorded along with their dates.
The data will eventually land in clickhouse and I want to be able to query it for things like how many a certain song has been played. or what the longest version of a specific song is.